+++ actionmailbox/CHANGELOG.md
@@ -1,3 +1,8 @@
+*   Add Amazon SES/SNS ingress.
+
+    *Bob Farrell*
+    *Chris Ortman*
+